+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Make `awless web` command no hidden anymore. This command allows to have a basic overview from a browser of the resources and theirs relations
 - Sync works on best effort now. Meaning it does not bail out when an error happens (most often it can be an access right issues on some AWS services)
 - Better template TAB completion: suggest on properties, suggest nothing if not relevant
+- Use `--no-headers` flag in `awless list` to display the results without headers
 - `awless ls policies` now returns: your managed policies + all policies attached to any users, role or group
 - New flag `--values-for` in `awless show` to output machine readable values for resource properties. Ex: `awless show my_instance --values-for name,publicip`
 - Listing [ECR](https://aws.amazon.com/ecr/) repositories: `awless list repositories`